Move implementation stuff out of `InMemoryPackResources`.
